+++ src/core/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@adcio.js/core",
+  "main": "./index.ts",
+  "dependencies": {
+    "@adcio.js/api": "workspace:*",
+    "@adcio.js/lib": "workspace:*"
+  }
+}
